Fix two known ASan-flagged bugs: find_any_record_var() and OP_SHL
Both surfaced during the virtual dispatch work and have shown up on
every ASan/UBSan sweep since (units, inherited, ParamCount/ParamStr),
logged in docs/ROADMAP.md's "Known issues" but not fixed until now.

find_any_record_var() (parser.c) read scope_record_var_count[-1] - a
global-buffer-overflow - whenever called at the top level (main
program body, nesting_depth == -1). Fixed with the same
nesting_depth >= 0 guard find_local() already had for the equivalent
plain-local lookup. Found and fixed the identical latent defect in
local_record_name_collides() while in the area, not yet observed to
be reachable in practice but the same unguarded read either way.

OP_SHL (vm.c) computed a << b directly on a signed int, which is
undefined behavior in C whenever the shifted-in bit doesn't fit back
in a signed int (b == 31 for any nonzero a). Fixed by shifting the
same bit pattern as unsigned int instead - always well-defined, same
two's-complement result this code already relied on in practice.

Both verified against their original ASan repros and a full
examples/ sweep with zero remaining sanitizer hits.

+++ src/solvm/vm.c
@@ -1369,7 +1369,14 @@ void run_vm(void) {
                     fprintf(stderr, "VM Runtime Error: Shift amount %d out of range (0..31)\n", b);
                     fatal_abort();
                 }
-                vm_push(&sp, a << b);
+                // Shifting a signed int left is undefined behavior in C
+                // whenever the result (or the bit shifted into the sign
+                // bit) doesn't fit back in a signed int - notably b == 31
+                // for any nonzero a. Shifting the same bit pattern as an
+                // unsigned int instead is always well-defined and yields
+                // the identical two's-complement result this code already
+                // relied on in practice, just without the UB.
+                vm_push(&sp, (int)((unsigned int)a << b));
                 break;
             }
